feat: implement splash screen
Implement the splash screen using the new `SplashScreen` API [1].
Set the start destination of the navigation graph to the splash screen.
Add a basic view model for the splash screen that simulates a login check with a delay and navigates to the home screen.
Add the splash screen module as a dependency in the navigation module.